feat: added tabs component
