+++ packages/UI/package.json
@@ -98,6 +98,7 @@
     "storybook": "7.2.2",
     "styled-components": "^5.2.1",
     "ts-jest": "28.0.7",
+    "type-fest": "4.3.1",
     "typescript": "^4.7.4"
   }
 }
